Atualizacao

+++ SimulaParcela.Dominio/Command/RegistrarNovaSimulacaoCommand.cs
@@ -5,11 +5,11 @@ namespace SimulaParcela.Dominio.Command
     public class RegistrarNovaSimulacaoCommand
     {
         public decimal ValorTotalCompra { get; private set; }
-        public decimal ValorJuros { get; private set; }
+        public int ValorJuros { get; private set; }
         public int QuantidadeDeParcela { get; private set; }
         public DateTime DataDaCompra { get; private set; }
 
-        public RegistrarNovaSimulacaoCommand(decimal valorTotalCompra, decimal valorJuros, int quantidadeDeParcela, DateTime dataDaCompra)
+        public RegistrarNovaSimulacaoCommand(decimal valorTotalCompra, int valorJuros, int quantidadeDeParcela, DateTime dataDaCompra)
         {
             ValorJuros = valorJuros;
             ValorTotalCompra = valorTotalCompra;
